Fix bug with SLAU and central/upwind blending
SU2 checks the upwind scheme to make sure that the central/upwind
blending is only used with supported numerical schemes. Previously,
SLAU and SLAU2 numerics had the capability to use central/upwind
blending of convective fluxes. However, the solver classes only
calculated the `Roe_Dissipation` for the Roe scheme. This check was also
duplicated across several functions.

This commit moves these checks of the upwind scheme to a single check in
the Config file preprocessing.  The SLAU and SLAU2 schemes are also
added to this check.