Unify the NTS and FD blending schemes
This commit changes the FD blending function to be consistent with both
published literature and the NTS blending. Before, Roe_Dissipation = 0
meant an upwind scheme for the FD blending function but a central scheme
for the NTS blending function, and vice versa for Roe_Dissipation = 1.
This commit changes the FD blending function to match the paper by
Molina et al. and the NTS blending scheme.

Since these two schemes now match, some of the code was also refactored
to eliminate duplication and improve readability.